Move package from internal/configprovider to service/internal/configprovider
