+++ service/config_provider.go
@@ -26,7 +26,7 @@ import (
 	"go.opentelemetry.io/collector/config/configmapprovider"
 	"go.opentelemetry.io/collector/config/configunmarshaler"
 	"go.opentelemetry.io/collector/config/experimental/configsource"
-	"go.opentelemetry.io/collector/internal/configprovider"
+	"go.opentelemetry.io/collector/service/internal/configprovider"
 )
 
 // ConfigProvider provides the service configuration.
