+++ service/internal/configprovider/merge.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package configprovider // import "go.opentelemetry.io/collector/internal/configprovider"
+package configprovider // import "go.opentelemetry.io/collector/service/internal/configprovider"
 
 import (
 	"context"
