+++ service/internal/configprovider/testdata/default-config.yaml
