+++ service/internal/configprovider/testdata/expand-with-all-env.yaml
