+++ service/internal/configprovider/testdata/expand-with-no-env.yaml
